feat(app): SettingsHubScreen wired at RouteNames.settings + PlaylistSourceSheet overflow fix
